+++ setup.py
@@ -69,6 +69,7 @@ if __name__ == "__main__":
         author_email="todo@fb.com",
         long_description="FairScale is a PyTorch extension library for high performance and large scale training on one or multiple machines/nodes. This library extends basic PyTorch capabilities while adding new experimental ones.",
         long_description_content_type="text/markdown",
+        entry_points={"console_scripts": ["wgit = experimental.wgit.__main__:main"]},
         classifiers=[
             "Programming Language :: Python :: 3.7",
             "Programming Language :: Python :: 3.8",
