+++ tests/ci_test_list_3.txt
@@ -23,3 +23,4 @@ tests/optim/test_oss_adascale.py
 tests/optim/test_ddp_adascale.py
 tests/experimental/nn/data_parallel/test_gossip.py
 tests/nn/data_parallel/test_fsdp_hf_transformer_eval.py
+tests/experimental/wgit/cli/test_cli.py
