+++ tests/experimental/wgit/cli/__init__.py
@@ -2,7 +2,3 @@
 #
 # This source code is licensed under the BSD license found in the
 # LICENSE file in the root directory of this source tree.
-
-from .cli import main
-
-__version__ = "0.0.1"
